+++ lib/fortitude/extensions/fortitude_ruby_ext.rb
@@ -2,10 +2,21 @@ require 'erb'
 
 ::String.class_eval do
   def fortitude_append_escaped_string(output)
+    _fortitude_append_escaped_string(output, false)
+  end
+
+  TABLE_FOR_ESCAPE_ATTRIBUTE_VALUE__ = {
+    '&' => '&amp;',
+    '"' => '&quot;'
+  }
+
+  def _fortitude_append_escaped_string(output, for_attribute_value)
     raise ArgumentError, "You can only append to a String" unless output.kind_of?(String)
 
     if html_safe?
       output.original_concat(self)
+    elsif for_attribute_value
+      output.original_concat(self.gsub(/[&\"]/, TABLE_FOR_ESCAPE_ATTRIBUTE_VALUE__))
     else
       output.original_concat(ERB::Util.html_escape(self))
     end
@@ -33,8 +44,8 @@ end
     each do |key, value|
       if value.kind_of?(Hash)
         new_prefix = case prefix
-        when String then fortitude_append_to(key, prefix.dup)
-        when nil then fortitude_append_to(key, "".html_safe)
+        when String then fortitude_append_to(key, prefix.dup, false)
+        when nil then fortitude_append_to(key, "".html_safe, false)
         else raise ArgumentError, "You can only use a String as a prefix"
         end
 
@@ -52,19 +63,19 @@ end
           else raise ArgumentError, "You can only use a String as a prefix"
           end
 
-          fortitude_append_to(key, target)
+          fortitude_append_to(key, target, false)
 
           if value == true
             if allows_bare_attributes
               # nothing here
             else
               target.original_concat(::Hash::FORTITUDE_EQUALS_QUOTE)
-              fortitude_append_to(key, target)
+              fortitude_append_to(key, target, false)
               target.original_concat(::Hash::FORTITUDE_QUOTE)
             end
           else
             target.original_concat(::Hash::FORTITUDE_EQUALS_QUOTE)
-            fortitude_append_to(value, target)
+            fortitude_append_to(value, target, true)
             target.original_concat(::Hash::FORTITUDE_QUOTE)
           end
         end
@@ -75,17 +86,17 @@ end
   end
 
   private
-  def fortitude_append_to(object, output)
+  def fortitude_append_to(object, output, for_attribute_value)
     case object
-    when String then object.fortitude_append_escaped_string(output)
-    when Symbol then object.to_s.fortitude_append_escaped_string(output)
+    when String then object._fortitude_append_escaped_string(output, for_attribute_value)
+    when Symbol then object.to_s._fortitude_append_escaped_string(output, for_attribute_value)
     when Array then object.each_with_index do |o,i|
       output.original_concat(" ") if i > 0
-      fortitude_append_to(o, output)
+      fortitude_append_to(o, output, for_attribute_value)
     end
     when nil then nil
     when Integer then output.original_concat(object.to_s)
-    else object.to_s.fortitude_append_escaped_string(output)
+    else object.to_s._fortitude_append_escaped_string(output, for_attribute_value)
     end
   end
 end