+++ spec/system/escaping_system_spec.rb
@@ -23,8 +23,16 @@ describe "Fortitude escaping behavior", :type => :system do
     expect(render(widget_class_with_content { p 'a<b' => 123 })).to eq("<p a&lt;b=\"123\"></p>")
   end
 
-  it "should escape attribute values" do
-    expect(render(widget_class_with_content { p :foo => 'a<b' })).to eq("<p foo=\"a&lt;b\"></p>")
+  it "should escape double quotes inside attribute values" do
+    expect(render(widget_class_with_content { p :foo => 'a"b' })).to eq("<p foo=\"a&quot;b\"></p>")
+  end
+
+  it "should escape ampersands inside attribute values" do
+    expect(render(widget_class_with_content { p :foo => 'a&b' })).to eq("<p foo=\"a&amp;b\"></p>")
+  end
+
+  it "should not escape less than signs, greater than signs, or single quotes inside attribute values" do
+    expect(render(widget_class_with_content { p :foo => 'a<b>c\'d' })).to eq("<p foo=\"a<b>c'd\"></p>")
   end
 
   it "should escape direct arguments to tags" do